Tests now convert lambda outputs and CPS outputs.
Instead of expecting the interface of both `Interp.interp`
and `Evalcps.interp` to have the same outputs, testing now
converts the output types to a new specialised output type
so that outputs across the two can be compared.

This special output type `eval` only supports a subset of the
lambda and CPS outputs and the tests must be made with this in mind.
For instance, it will not accept a lambda or CPS output that is a
function.

Most changes here are within the testing library alone, but some
small changes in `Interp` were necessary. I changed `Interp.interp`
to output an answer directly and then `interp.mli` to export the
answer type so we can match on it in the testing library.